+++ product_country_restriction/readme/HISTORY.rst
@@ -2,3 +2,11 @@
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 * First version
+
+10.0.1.1.0 (2019-01-14)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Allows to define different strategies by company:
+* * Authorize (default): Authorize all products except those with a
+    rule
+* * Restrict: Restrict all products except those with a rule